re organize logs for migrations to be correctly logged
