+++ api/config/config.js
@@ -1,3 +1,12 @@
+let winston = require('winston');
+let logging = require('../util/logging');
+
+const log_level = process.VERBOSITY || (new Set(["test", "development"]).has(process.env.NODE_ENV) ? "debug" : "info");
+let db_logger = winston.loggers.add("db", {
+  transports: [ new winston.transports.Console() ],
+  level: log_level,
+  format: logging.get_default_format("DB")
+});
 
 let database = {
     username: process.env.DB_USERNAME,
@@ -8,7 +17,7 @@ let database = {
     timezone: process.env.TIMEZONE || "UTC",
     seederStorage: "sequelize",
     logging: (msg) => {
-      require("winston").loggers.get("db").debug(msg);
+      db_logger.debug(msg);
     }
 };
 
@@ -26,6 +35,10 @@ module.exports = {
     email_settings: {
         sender_name: 'BoardGameComponion',
         email_address: 'info@boardgamecomponion.com'
+    },
+    log_level,
+    loggers: {
+      db: db_logger
     },
     sendgrid_api_key: process.env.SENDGRID_API_KEY,
     [process.env.NODE_ENV]: database
