+++ api/util/logging.js
@@ -20,3 +20,22 @@ exports.logError = (logger, err) => {
   logger.error(err_info.error);
   logger.error(err_info.stack);
 };
+
+// define general log format
+const log_format = winston.format.printf((options) => {
+  return `[${options.timestamp}] [${options.level}] [${options.label}] ${options.message}`;
+});
+
+/**
+ * Create a general format winston object for logs
+ * @param label A logger identifier to be printed in the string
+ * @returns {Format}
+ */
+exports.get_default_format = (label) => {
+  return winston.format.combine(
+      winston.format.label({label}),
+      winston.format.colorize(),
+      winston.format.timestamp(),
+      log_format
+  );
+};
